Switched to floats, removed unused code. Cleaned up unit tests.
commit 5f3e06b9836557c30008a6463fd0eee5660a7b43
Author: Robert C. Taylor <r0wbrt@gmail.com>
Date:   Mon Jun 26 17:31:42 2017 -0500

    Hlint and format unit tests

commit 90bb4db5f27acc2556801ec7e9546fbd24b170dc
Author: Robert C. Taylor <r0wbrt@gmail.com>
Date:   Mon Jun 26 17:10:46 2017 -0500

    Fix unit tests float comparison

https://randomascii.wordpress.com/2012/02/25/comparing-floating-point-numbers-2012-edition/

commit df8ed2f32e2bb8a09368bf9682e3c6c7ee8e2ece
Author: Robert C. Taylor <r0wbrt@gmail.com>
Date:   Mon Jun 26 13:21:00 2017 -0500

    Do not do -O2 by default. Update Version

commit 7187194fd0f871b7f73d9d8cbe72650c08a3905e
Author: Robert C. Taylor <r0wbrt@gmail.com>
Date:   Mon Jun 26 13:18:18 2017 -0500

    Remove Old Tests, Delete OFDM support (Issue 19)

commit eccd59c12cfde0ba1f0f4c0fce6f573289c05ea7
Author: Robert C. Taylor <r0wbrt@gmail.com>
Date:   Sun Jun 18 14:56:26 2017 -0500

    Switched to floats

+++ RadarTools/src/lib/YASDRR/DSP/Correlation.c
@@ -25,23 +25,23 @@ limitations under the License.
  * the signal array in size.
  */
 __attribute__((hot))
-void complexCorrelate(int kSize, int sSize, double * kernel, double * signal, double * output) 
+void complexCorrelate(int kSize, int sSize, float * kernel, float * signal, float * output) 
 {
     int i;
     int j;
-    double Raccumulator;
-    double Iaccumulator;
+    float Raccumulator;
+    float Iaccumulator;
     for(i = 0; i < sSize; i++) 
     {
         Raccumulator = 0;
         Iaccumulator = 0;
         for(j = 0; j < kSize && (j+i) < sSize; j++)
         {
-            double a = kernel[2*j];
-            double b = kernel[2*j + 1];
+            float a = kernel[2*j];
+            float b = kernel[2*j + 1];
             
-            double c = signal[2*i + 2*j];
-            double d = signal[2*i + 2*j + 1];
+            float c = signal[2*i + 2*j];
+            float d = signal[2*i + 2*j + 1];
             
             //Signs are flipped because the kernel is conjugated.
             Raccumulator += a * c  + b * d;
